// Source/DocuEye.Web/client-app/app/api/docueye-api/models/element.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DocuEye.Web
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export interface Element {
    'id'?: string | null;
    'dslId'?: string | null;
    'name'?: string | null;
    'url'?: string | null;
    'group'?: string | null;
    'tags'?: Array<string> | null;
    'type'?: string | null;
    'location'?: string | null;
    'technology'?: string | null;
    'description'?: string | null;
    'parentId'?: string | null;
    'instanceOfId'?: string | null;
    'workspaceId'?: string | null;
    'properties'?: { [key: string]: string; } | null;
    'sourceCodeUrl'?: string | null;
    'ownerTeam'?: string | null;
    'structurizrId'?: string | null;
}